refactor(cli): move account options from global options
