+++ packages/tools/kadena-cli/src/account/commands/accountAddManual.ts
@@ -2,6 +2,7 @@ import { assertCommandError } from '../../utils/command.util.js';
 import { createCommandFlexible } from '../../utils/createCommandFlexible.js';
 import { globalOptions } from '../../utils/globalOptions.js';
 import { log } from '../../utils/logger.js';
+import { accountOptions } from '../accountOptions.js';
 import { addAccount } from '../utils/addAccount.js';
 import { displayAddAccountSuccess } from '../utils/addHelpers.js';
 import { getAccountDetails } from '../utils/getAccountDetails.js';
@@ -11,12 +12,12 @@ export const createAddAccountManualCommand = createCommandFlexible(
   'add-manual',
   'Add an existing account locally to the CLI',
   [
-    globalOptions.accountAlias(),
-    globalOptions.accountName(),
+    accountOptions.accountAlias(),
+    accountOptions.accountName(),
     globalOptions.fungible(),
     globalOptions.networkSelect(),
     globalOptions.chainId(),
-    globalOptions.accountOverwrite(),
+    accountOptions.accountOverwrite(),
     globalOptions.publicKeys(),
     globalOptions.predicate(),
   ],
