+++ packages/tools/kadena-cli/src/account/commands/accountResolveAddressToName.ts
@@ -3,6 +3,7 @@ import { assertCommandError } from '../../utils/command.util.js';
 import { createCommandFlexible } from '../../utils/createCommandFlexible.js';
 import { globalOptions } from '../../utils/globalOptions.js';
 import { log } from '../../utils/logger.js';
+import { accountOptions } from '../accountOptions.js';
 import { kdnResolveAddressToName } from '../utils/txKdnResolverChain.js';
 
 export const resolveAddressToName = async (
@@ -40,7 +41,7 @@ export const resolveAddressToNameCommand = createCommandFlexible(
   'Resolve a k:address to a .kda name (kadenanames)',
   [
     globalOptions.network({ isOptional: false }),
-    globalOptions.accountKdnAddress(),
+    accountOptions.accountKdnAddress(),
   ],
   async (option) => {
     const kadena = await option.network({
